refactor: move leverage type to types
